Refactored mthod/function names.
To make it easier to make out the difference between method names that give the ability to manipulate the peripheral and the ones that update the delegate with new informations / characteritic values, I decided to refactor the method names.

Also added an optional delegate call for battery charging state and dcdc usage.

All calls to the delegate now are prefixed with cc_did...